JBDS-3996 Build the installer jar in package phase
Until now, the installer jar was only built in the installer phase.
That's because the izpack ant call had install phase set.
I think it makes sense to do this in package phase already.
From maven.apache.org:
install - install the package into the local repository, for use as a dependency in other projects locally